Space-Pin-Pong/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==29.0.0
filelock==3.15.4
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
identify==2.6.0
idna==3.7
incremental==24.7.2
Markdown==3.6
msgpack==1.0.8
nodeenv==1.9.1
numpy==2.1.0
overrides==7.7.0
pip==24.2
platformdirs==4.2.2
pre-commit==3.8.0
priority==1.3.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wheel==0.43.0
zope.interface==7.0.3
